dockerstart.sh

+++ dockerstart.sh
@@ -1,7 +1,18 @@
 #!/bin/bash -l
+: ${BRANCH=master}
+
 cd /meanshop
-git reset --hard HEAD
-git pull http origin
+if [ -d ".git" ]; then
+    git reset --hard HEAD
+    git fetch --prune http
+    git checkout $BRANCH
+else
+    git init
+    git remote add http https://github.com/Thalhalla/meanshop.git
+    git fetch --prune http
+    git checkout $BRANCH
+fi
+
 npm install
 bower install
 grunt serve
